resolve: remove runtime fixture and unify profile auto-pick in ai-doc route
